Alan ... fijate si podes codear solamente el loop e run game
    El set Snap falta tambien pero veo si lo hago yo

+++ server/source/AcceptThread.cpp
@@ -7,7 +7,7 @@ void AcceptThread::disconect() { menu.disconectAll(); }
 void AcceptThread::run() {
   try {
     Menu menu;
-    int a=0;
+    size_t a=0;
     while (serverIsOpen || a!=2) {
       Socket sktAccepted = soktRef.accept();
       if (serverIsOpen) menu.conectNewClient(std::move(sktAccepted),a);
